ci: update checkout action
github is listing warnings that indicate that the versions of 'actions'
that i'm using in CI need updating due to use of obsolete features.

i found the github repo for the 'checkout' action and a new version was
available, so i'm updating to that here.

i also found what i believe is the repo for the 'actions-rs/toolchain'
action, but it seems that version 1 is still the latest.