add `set_corked_state()` method to `Stream`
the C API just has a single function to toggle state. when i first built
this binding i thought that separate `cork()` and `uncork()` methods might
be much nicer. however for some users it may hypothetically work better to
just be able to submit a boolean state to a single function, as with the
C API. additionally as it was we had almost identical duplicated code for
each function.

adding this new function de-duplicates the code and lets the user choose
which solution works best for them.