comment out compile-fail test
the `cfg(compile_fail)` attribute relied upon 'features', requiring the
following test command: `cargo test --features compile_fail`. this was not
being covered in automated testing. since rust 1.80 when running `cargo
test` i get a warning output about `compile_fail` being an unknown config
attribute (not included in the list of features in the cargo.toml file).
if i try to execute it with the feature flag enabled i get an error
preventing me from using it.

there is no `#[compile_fail]` attribute yet, rust only supports this in
doc comment tests.

so solutions are either to:
 - move this to a doc test, but i don't want this to appear in
   documentation.
 - add `compile_fail` to the list of features
 - comment out the code and only uncomment it if/when wanting to check that
   the test still works.

if i decide to always include it in automated testing then i'll have to add
it to the features list. otherwise i'll just comment it out for now.