ensure proplist `IntoIterator` impl is only used with owned instances
it output's an instance with a static lifetime, which we would not want to
do with a borrowed proplist instance. luckily the only possible
free-standing proplist instances are owned ones created via the `new()`
functions. all borrowed instances are only accessible via a reference and
thus cannot be used with `into_iter()` which comsumes the input.